refactor(fetch): move fetch functions in one file
